Finally migrate everything to accounts v3
This has been on the cards for ages, but now finally lets us drop the
auth0 dep completely, brings the UI in line with the website & accounts
dashboard, and sets us up for cloud integrations for public endpoint
shortly. Integration with the new k8s-based accounts systems is tested
but not yet activated by this, we'll need to set that up as well
shortly, with various other changes there en route.